day30/autoquant-day30/requirements.txt
alpaca-py>=0.13.0
numpy>=1.26.0
pandas>=2.1.0
rich>=13.7.0
python-dotenv>=1.0.0
pytest>=7.4.0
